Make A Vendor C
No outwardly observable behaviour has changed. This is because the new
version of C which was being called in the GOPATH by A, is now vendored
into A. However, if C changes in the GOPATH, we'll see some effect,
because we'll continue to use the vendored copy from A, not the changed
copy.