dep init ; dep ensure
This vendors B and C for us, inside A's vendor directory. It's clear why
we get the version of B that we get -- it's the only one. However, we
were using version 2 of C, while B was using version 1 of C. Dep
vendored version 1 of C for us. Why was this?

Some possible reasons:
- it gave us the version of C that B is using
- it gave us the default version of C as if we did a go-get

In any case, we now have a single consistent version of C -- we only
have version 1. If we `go run src/a/main.go`, we only get badgers
printed, and no ostriches. If we
`go test src/a/vendor/github.com/totherme/govendorexplorationb`, the
tests pass.