dep init ; dep ensure
This seems to have ignored the branch of B that we were actually using
(in the $GOPATH), and just gotten the master branch from github. The
result is that we have version 1 of everything. `go run src/a/main.go`
prints only badgers, and
`go test src/a/vendor/github.com/totherme/govendorexplorationb` passes
-- using version 1 (the badgers version) of C.