Use the deliberately vague version of B
In the B submodule (in the `$GOPATH`), we create a deliberately vague
branch. In this branch, we edit the Gopkg.toml to remove all constraints
on which version of C we require. Thus, we allow A to use whichever
version it wants. To test this, we constrain A to version 1 of C. `dep
ensure` is happy with this.