Check that C is a transitive dep, even when vague
Write Gopkg.toml by hand, to require only the deliberately vague version
of B. We know for sure that this version of B doesn't explicitly require
C in its Gopkg.toml. Will this cause our program to fail to vendor C as
a transitive dependency?

The answer is: no -- everything is fine.

When we run `dep ensure`, dep discovers the transitive dependencies,
even though no explicit constraints are declared. The result is that a
sensible default C is vendored in (whatever the default branch is -- in
this case version 1)