godep save -t
This vendors B and C for us, inside A's vendor directory. It's clear why
we get the version of B that we get -- it's the only one. However, we
were using version 2 of C, while B was using version 1 of C. Dep
vendored version 2 of C for us. Why was this?

Some possible reasons:
- it gave us the version of C that's checked out in the `$GOPATH`

We now have a version of A which is consistent as it uses only version 2 of C.
When we run `go run src/a/main.go`, we only get ostriches, no badgers.

However, `go test ./src/a/vendor/github.com/govenderexplorationb` fails. As C's
version from the `$GOPATH`, version 2, was vendored in, we have a missmatch
between the tests in B -- they still expect badgers -- and the actual version
of C which now prints ostriches.